Module Geometry.MultiPoint

type t

A multipoint is an array of positions.

val coordinates : t -> Position.t array

Get the positions that make up this multipoint object.

val v : Position.t array -> t

Create a multipoint object from an array of positions.

val of_json : json -> (t, [ `Msg of string ]) Stdlib.result

val to_json : t -> json